A theory that I have, which is mine, which is to say Dinosaurs are...
Lot's o' refactoring.  Eliminated a lot of the state in favor of
functional derivation, as the mutlicolumn algorithm requires
reevaluating vertical layout with each shift.  As the max label width
can change with each shift, increasing or decreasing the element height
of the column, we move to a more elegant functional system.  The
resultant control is the stateful layout result of the system.