whew.  Fixed table vs outline layout.
I needed to:

a) recalculate the outline width, as the nesting of tables below
obviously changes this calculation.
b) compare the table width with the newly calculated outline width.

I was incorrectly thinking I needed to calculate the available space
from the supplied justified width subtracting the calculated label width
of the relation being laid out.

My error was not realizing the passed in justified width is the complete
width given to this relation.  So, no subtraction.