ugh.  finally got that niggling extraction back in place.
Things look 5x5, finally.  Smooth as silk, without the Control
framework.  Don't know why that flashes so much, but it's ugly.  Way
nicer this way.  And far lighter weight.  Will have to add the behavior,
which kind of sucks, but the well behaved framework looks like it will
be quite nice to work with.