Huge refactoring.  Some semblence of sanity.
Finally bit the bullet and fully excised the SchemaNode from the layout,
after the layout is initially built in measurement.  Folds are handled
in the measurement stage.  The resulting schema node layout of the
measurement is now the fully laid out structure.  No need to bounce
around the schema node structure in a visitor pattern.

Geebus.  That was a long time coming.

Still not handling the fold content correctly, but needed a checkpoint
after all that.